backend/requirements.txt
Django==5.0.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
python-dotenv==1.0.1
dj-database-url==2.1.0
psycopg2-binary==2.9.9  # PostgreSQL adapter, remove if using SQLite
pillow==10.2.0